feature: update and delete accommodation

+++ impl/src/main/java/com/grimels/lazurcity/mapper/AccommodationMapper.java
@@ -23,7 +23,7 @@ public interface AccommodationMapper {
         entity.setRoom(roomEntity);
         entity.setStartDate(createAccommodationRequest.getStartDate());
         entity.setEndDate(createAccommodationRequest.getEndDate());
-        entity.setPrice(createAccommodationRequest.getPrice());
+        entity.setPrice(createAccommodationRequest.getPrice().doubleValue());
         entity.setQuantity(createAccommodationRequest.getQuantity());
 
         return entity;
